1/28任务全部完成🧨(昨天去吃年饭，忘了push！)

+++ 刷题笔记/2022/【入门3】循环结构/P5723 【深基4.例13】质数口袋.cpp
@@ -39,7 +39,7 @@ int isPrime(int n){
 
 int main(){
 	int n;
-	int ans = 0, count = 1;
+	long long ans = 0, count = 1;
 	cin >> n;
 
 	if(n < 2){
